Move the reservation and loging out of exec module
The lr/sc reservation was moved out of the exec module. The amo* instructions  locking is enabled.

Address tag bits are used to implement the registration/invalidation of the register set for lr/sc instructions and for amo* instructions memory locking.

The actual reservation and locking is to be implemented.