Use non-Wishbone interface for faster pipeline
The decoder, regfile and exec modules no longer use a Wishbone interface. The cyc signal was removed and the operation that each modules executes starts with a pulse on stb from the master and ends with an ack pulse from the slave. This allows the master to start a new transaction in the same cycle as the one in which the current one ends. The pipeline does exactly that when the decoder, the regfile and the exec complete a transaction. A slight performance improvement is registered due to a faster pipeline.